Server/requirements.txt
pyserial
tinydb
rpigpio